+++ src/main/java/cart/dao/CartDao.java
@@ -1,7 +1,13 @@
 package cart.dao;
 
+import cart.entity.Cart;
+
+import java.util.List;
+
 public interface CartDao {
     int insert(int productId, String email);
 
     int delete(int cartId, String email);
+
+    List<Cart> selectById(String email);
 }
