--- src/main/java/cart/dao/CartProductDao.java
@@ -1,9 +1,0 @@
-package cart.dao;
-
-import cart.entity.Cart;
-
-import java.util.List;
-
-public interface CartProductDao {
-    List<Cart> selectById(String email);
-}
